+++ docs/legacy_sanity_check/CHANGELOG.md
@@ -2,7 +2,12 @@
 
 All notable changes to this gear are documented in this file.
 
+## 0.1.3
+* Skips sanity checks for affiliated studies
+  
+## 0.1.1
+* Skips sanity checks for inactive centers
+  
 ## 0.1.0
-
 * Initial version
 * Adds this CHANGELOG